fix: changed few datatypes

+++ backend/src/types/mint.rs
@@ -4,10 +4,10 @@ use serde::{Deserialize, Serialize};
 pub struct MintData{
     pub mint_address : String,
     pub owner : String,
-    pub data_length : u64,
+    pub data_length : usize,
     pub mint_authority : String,
-    pub supply : f64,
-    pub decimal : u32,
+    pub supply : u64,
+    pub decimal : u8,
     pub is_initialized : bool,
     pub freeze_authority : Option<String>
 }
